requirements.txt
selenium
aiohttp
asyncio
python-telegram-bot --pre
python-dotenv
pillow
bs4
